backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "google-genai==0.8.0",
    "google-generativeai>=0.8.5",
    "pydantic==2.9.0",
    "python-dotenv==1.0.0",
    "uvicorn[standard]==0.31.0",
]
